requirements.txt
# Dependências do projeto
Flask==2.3.3
mysql-connector-python==8.2.0
python-dotenv==1.0.0

# Dependências opcionais para desenvolvimento
pytest==7.4.3
pytest-flask==1.2.0
black==23.9.1
flake8==6.1.0

# Dependências para produção
gunicorn==21.2.0
waitress==2.1.2